Initial CPU debug info

+++ kernel/include/cpuid.h
@@ -19,7 +19,7 @@ extern const char __cpuid_vendor[13];
 
 void init_cpuid(void);
 
-bool cpuid(uint8_t leaf, uint32_t *eax, uint32_t *ebx, uint32_t *ecx,
+bool cpuid(uint32_t leaf, uint32_t *eax, uint32_t *ebx, uint32_t *ecx,
            uint32_t *edx);
 
 #endif //__ANOS_KERNEL_CPUID_H
